# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Aggregato API
  version: 1.0.0
  description: |
    Single-user, self-hosted media log aggregator. Every endpoint requires authentication by
    default: either `Authorization: Bearer <api.token>` or the session cookie issued by
    `POST /auth/session`. When `api.allow_unauthenticated_readonly` is enabled, credential-free
    GET/HEAD requests are allowed as read-only requests; writes and websockets still require a
    credential. Backups, retained diagnostics, and ingest-failure payloads require operator
    access even though they use GET. Public and read-only credentials receive 403 on those routes.

    An optional second credential, `api.readonly_token`, authenticates reads only: any
    POST/PUT/PATCH/DELETE it presents — directly or through the session cookie it exchanges
    itself for — is refused with 403 `readonly-token`. `GET /auth/session` reports which
    credential the caller holds.

    Conventions:
      * Timestamps are ISO 8601 with offset. `logged_at` is always accompanied by
        `logged_precision`; clients must not render a fabricated time as exact.
      * Ratings return raw value, scale, and normalized value. Normalized values are
        comparable within a scale, not across platforms.
      * Pagination is keyset-only via opaque `cursor`. There is no offset parameter.
      * Collection pages use `limit` 1–200, defaulting to 50; out-of-range values are clamped.
...
        is no arbitrary matching-result cap.
...
servers:
...
security:
  - bearerAuth: []
  - sessionCookie: []
...
paths:
  /auth/session:
    post:
      tags: [operations]
      summary: Exchange the API token for a session cookie
      description: Sets an HttpOnly, SameSite=Lax cookie and returns a CSRF token for
        cookie-authenticated writes. Rotating `api.token` invalidates all sessions.
        Accepts `api.readonly_token` too, which yields a read-only session.
      security: [{ bearerAuth: [] }]
      responses:
        "204":
          description: Session established; cookie and CSRF token set
          headers:
            Set-Cookie: { schema: { type: string } }
        "401": { $ref: "#/components/responses/Problem" }
    get:
      tags: [operations]
      summary: Describe the current credential
      description: What the caller is authenticated as. The web UI reads `readonly` to hide
        every control it would only be refused for.
      responses:
        "200":
          description: The current credential
          content:
            application/json:
              schema:
                type: object
                required: [via, readonly]
                properties:
                  via: { type: string, enum: [bearer, cookie, public] }
                  readonly: { type: boolean }
        "401": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
    sessionCookie: { type: apiKey, in: cookie, name: aggregato_session }
...
  responses:
    Problem:
      description: RFC 9457 problem detail
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...